[AutoDiff] Enables conditional differentiability on protocol requirements. (#25974)
This PR enables the functionality shown in this example:

```swift
public protocol Distribution {
  associatedtype Value
  func logProbability(of value: Value) -> Float
}

public protocol DifferentiableDistribution: Differentiable, Distribution {
  @differentiable(wrt: self)
  func logProbability(of value: Value) -> Float
}

struct Foo: DifferentiableDistribution {
  @differentiable(wrt: self)
  func logProbability(of value: Float) -> Float {
    .zero
  }
}

@differentiable
func blah<T: DifferentiableDistribution>(_ x: T) -> Float where T.Value: AdditiveArithmetic {
  x.logProbability(of: .zero)
}
```

The fix is based on the fact that the Swift compiler does not add entries to the witness tables of protocols for overridden functions, to avoid redundancy. However, the `@differentiable` attribute being added should not be interpreted as an override as it adds new functionality, and should result in entries being added to the witness table. This PR adds this check to the override checking code, thus enabling support for the aforementioned feature.