AST: add a workaround for VS2017
Unfortunately, VS2017 does not support the SFINAE expression.  This
means that it attempts to parse the `Index < sizeof...(Types)` as a
template parameter resulting in a parse failure.  This normalises the
constraint to `0 < sizeof...(Types) - Index`
(`sizeof...(Types) - Index > 0`) which is then wrapped into a ternary to
explicitly convert it to a boolean.  This repairs the builds on VS2017.