[Sema] Allow unavailable extension of unavailable type in #if
The compiler allows unavailable types to be extended as long as the
extension itself is marked unavailable:

@available(macOS, unavailable)
struct UnavailableStruct { }

@available(macOS, unavailable)
extension UnavailableStruct { } // no-error

However, this doesn't work when the extension is inside a #if. The
availability logic that finds the containing ExtensionDecl based on
a source range gets confused and finds the IfConfigDecl instead.

To fix this, change that logic to ignore IfConfigDecls. Members of
an active #if are also represented in the enclosing context, as
sibling of the IfConfigDecl, so the compiler will find the extension
declaration there.

rdar://problem/53079366