IRGen: Don't get enum element argument type until we need it
Instead of adding a resolveDeclSignature() call here, I'm going to live
dangerously and try to only get the enum element type in the case where
SIL type lowering has already computed it, that is, if the enum is not
indirect.

Soon this will become moot anyway because getInterfaceType() will be a
request.