stdlib: Print the error message in precondition() and preconditionFailure() even in release builds.
This is important for frameworks, like Foundation, because frameworks are always shipped and used as release builds.
This change results in a little code size overhead, but it's quite small: Instead of a trap instruction it basically ends up in loading a few registers and calling the fatalError runtime function.

I also added an overload for precondition/Failure without a message (instead of a default parameter). This overload just traps, because without a message calling the fatalError function does not add any benefit.