Strengthen SystemRandomNumberGenerator docs.
SE-0202 claimed that it would add a secure RNG for all platforms, but
the documentation for SystemRandomNumberGenerator contains a few weasel
phrases that imply that the strength may not meet cryptographic
requirements on all platforms.

This language makes the tool worse than useless, as any unexpected
future platform may implement a weak random number generator. We should
make the documentation reflect the spirit of SE-0202 and have it clearly
state that all implementations will be suitable for cryptographic
applications.

While we're here, I added some documentation to cover the Windows
implementation, which was missed previously.