[presets] Add clang install to bots that test indexstore-db/sourcekit-lsp
At a minimum, we need to install libIndexStore to enable testing the
indexer, but it's better to also include clang/clangd, since this lets
us enable all of our tests on these bots. This commit uses the
swift-install-components and llvm-install-components from the Package
bots, which increases the similarity between our PR tests and package
bots.

In addition to the sourcekit-lsp/indexstore-db PR test bots, this
affects the following bots that test indexstore-db and didn't already
install clang:

* Linux
  * incremental
  * incremental long test
  * swiftpm PR test

* macOS
  * PR test
  * PR smoke test
  * incremental
  * swiftpm PR test

Based on looking at recent build logs for the package bots, this should
only add <5 seconds to the incremental build.