[IDE] Fix document structure request output for interpolated string literals
InterpolatedStringLiteralExpr has a TapExpr, which contains a BraceStmt
containing the builder CallExprs to the builder appendInterpolation /
appendStringLiteral methods used to construct the final string. This is all
implementation detail that doesn't actually appear in the source code, but the
document structure request didn't filter it out, resulting in bogus calls and
brace ranges in its responses.

Resolves rdar://problem/55183943