[android] Fix test setup for Android to avoid using -sdk.
The recently introduced changes to make the -sdk also the resource
directory broke the tests in Android, since -sdk was being used because
it is passed as --sysroot to child invocations, which was convinient.

This commit removes those usages in the test suite, and switches to use
--sysroot for the compiler and the linker. Even if in Android only the
linker bit is necessary (since the NDK headers are somewhere different,
and are passed as -I parameters to Swift).