ClangImporter: search for `glbc.modulemap` in the SDK
When cross-compiling with a swift built for another target or when
building with a pure toolchain without the resource directory containing
host content, the modulemap required for Linux may not be found.  In
such a case, we should also look for the content in the SDK if one is
provided.  This improves the behaviour of `swiftc` for Linux with a
standalone SDK.